use the new select helper methods in UVFlag
refactor some more code into utils methods to allow reuse
