+++ src/pyuvdata/utils/__init__.py
@@ -18,6 +18,7 @@ RADIAN_TOL = 1 * 2 * np.pi * 1e-3 / (60.0 * 60.0 * 360.0)
 LST_RAD_TOL = 2 * np.pi * 5e-3 / (86400.0)
 
 # these seem to be necessary for the installed package to access these submodules
+from . import antenna  # noqa
 from . import apply_uvflag  # noqa
 from . import array_collapse  # noqa
 from . import bls  # noqa
